SeqNumBytes moved to core
