+++ zenoh/src/net/transport/unicast/establishment/authenticator/userpassword.rs
@@ -16,7 +16,7 @@ use super::{
 };
 use super::{Locator, WBuf, ZBuf, ZInt, ZenohId};
 use crate::config::Config;
-use crate::net::protocol::message::SeqNumBytes;
+use crate::net::protocol::core::SeqNumBytes;
 use crate::net::transport::unicast::establishment::Cookie;
 use async_std::fs;
 use async_std::sync::{Arc, Mutex, RwLock};
